TransitFraud/requirements.txt
setuptools
dataclass_csv
google-cloud-spanner
Faker
Flask
flask_bootstrap
